GetTokenUseCase 통합 테스트 구현
